+++ internal-api/src/test/groovy/datadog/trace/bootstrap/instrumentation/ci/AppVeyorInfoTest.groovy
@@ -1,6 +1,6 @@
 package datadog.trace.bootstrap.instrumentation.ci
 
-class AppVeyorInfoTest extends CIProviderInfoTest {
+class AppVeyorInfoTest extends CITagsProviderImplTest {
 
   @Override
   CIProviderInfo instanceProvider() {
