+++ internal-api/src/test/groovy/datadog/trace/bootstrap/instrumentation/ci/BitBucketInfoTest.groovy
@@ -1,6 +1,6 @@
 package datadog.trace.bootstrap.instrumentation.ci
 
-class BitBucketInfoTest extends CIProviderInfoTest {
+class BitBucketInfoTest extends CITagsProviderImplTest {
 
   @Override
   CIProviderInfo instanceProvider() {
