+++ internal-api/src/test/groovy/datadog/trace/bootstrap/instrumentation/ci/BuddyInfoTest.groovy
@@ -1,6 +1,6 @@
 package datadog.trace.bootstrap.instrumentation.ci
 
-class BuddyInfoTest extends CIProviderInfoTest {
+class BuddyInfoTest extends CITagsProviderImplTest {
 
   @Override
   CIProviderInfo instanceProvider() {
