+++ internal-api/src/test/groovy/datadog/trace/bootstrap/instrumentation/ci/BuildkiteInfoTest.groovy
@@ -1,6 +1,6 @@
 package datadog.trace.bootstrap.instrumentation.ci
 
-class BuildkiteInfoTest extends CIProviderInfoTest {
+class BuildkiteInfoTest extends CITagsProviderImplTest {
 
   @Override
   CIProviderInfo instanceProvider() {
