+++ internal-api/src/test/groovy/datadog/trace/bootstrap/instrumentation/ci/CircleCIInfoTest.groovy
@@ -1,6 +1,6 @@
 package datadog.trace.bootstrap.instrumentation.ci
 
-class CircleCIInfoTest extends CIProviderInfoTest {
+class CircleCIInfoTest extends CITagsProviderImplTest {
 
   @Override
   CIProviderInfo instanceProvider() {
