+++ internal-api/src/test/groovy/datadog/trace/bootstrap/instrumentation/ci/JenkinsInfoTest.groovy
@@ -1,6 +1,6 @@
 package datadog.trace.bootstrap.instrumentation.ci
 
-class JenkinsInfoTest extends CIProviderInfoTest {
+class JenkinsInfoTest extends CITagsProviderImplTest {
 
   @Override
   CIProviderInfo instanceProvider() {
