+++ internal-api/src/test/groovy/datadog/trace/bootstrap/instrumentation/ci/TravisInfoTest.groovy
@@ -1,6 +1,6 @@
 package datadog.trace.bootstrap.instrumentation.ci
 
-class TravisInfoTest extends CIProviderInfoTest {
+class TravisInfoTest extends CITagsProviderImplTest {
 
   @Override
   CIProviderInfo instanceProvider() {
