+++ internal-api/src/test/groovy/datadog/trace/bootstrap/instrumentation/ci/UserSuppliedCIInfoTest.groovy
@@ -1,6 +1,6 @@
 package datadog.trace.bootstrap.instrumentation.ci
 
-class UserSuppliedCIInfoTest extends CIProviderInfoTest {
+class UserSuppliedCIInfoTest extends CITagsProviderImplTest {
 
   @Override
   CIProviderInfo instanceProvider() {
